+++ src/redirector.py
@@ -1,4 +1,4 @@
-import utils
+import utils.utils as utils
 
 
 class IORedirector(object):
